Sample rate (mostly) fixed
Inability to keep up was due to the modulo operation being computationally expensive (Thank you to build_1971 for spotting the problem). To make software trigger feasible, this fix requires ADCBUFFERSIZE to be a power of two so that ANDing the counter with ADCBUFFERSIZE-1 is equivalent to the modulo. (Thank you to aarg for the suggestion.) This works with a prescaler as small as 32 (153 kS/s) but the menu option allows for going faster for now in the hope I might eventually squeeze a little more out of it. Committing also sketch adctest, which demonstrates operation at maximum rate but without any fanciness like a software trigger. No longer using interrupts to collect data - not sure which approach is better. Not really using much Girino code at this point.